1801 - code review changes
