+++ scripts/google_dataset/schema_org_sparql.py
@@ -1,4 +1,4 @@
-# This script is responsible for generating a validation.json file that contains all attributes
+# This script is responsible for generating a jsonldSchema.json file that contains all attributes
 # and its types for the schema.org classes used as input
 # schemaorg-current-https.ttl is used as input for the data definition
 # Output can be compared against schema.org/CLASS_NAME
@@ -13,7 +13,7 @@ dataDownloadRequiredProps = ['contentUrl', 'encodingFormat']
 def transform(rowElement, attr):
     return rowElement[attr].toPython().split("/")[-1]
 
-
+# convert the rowElement to an array 
 def transformToArray(rowElement, attr):
     individualElements = rowElement[attr].toPython().split(" ")
     sortedElements = sorted(individualElements)
@@ -27,7 +27,8 @@ def transformToArray(rowElement, attr):
 g = rdflib.Graph()
 g.parse("schemaorg-current-https.ttl", format="n3")  # query the graph
 
-# TODO: the DISTINCT doesn't seem to work properly. Need a better way to aggregate.
+# Fetches all the rows and then filters based on the range type and domain type
+# We also filter the attributes that have been superseded by newer attributes in schema.org
 result1 = g.query("""
 SELECT 
   ?prop
@@ -74,6 +75,7 @@ GROUP BY ?prop
 ORDER BY ?prop
 """)
 
+# We get the parent class of all the classes with this query
 result2 = g.query("""
 SELECT 
   ?prop ?subclass
@@ -86,20 +88,18 @@ ORDER BY ?prop
 """)
 
 allSchemaOrgTypesDict = {}
-count = 0
 attrMap = {}
 for row in result1:
     domainArr = transformToArray(row, 'dtypes')
     label = transform(row, 'label')
     dataType = transformToArray(row, 'rtypes')
     for domain in domainArr:
+       # If schema.org class has not been seen before then setup the structure else just append
         if domain not in allSchemaOrgTypesDict:
             allSchemaOrgTypesDict[domain] = {"properties": {}, "requiredProperties": [], "parent": None}
 
         schemaClassFromDict = allSchemaOrgTypesDict[domain]
         schemaClassFromDict["properties"][label] = {"types": dataType}
-        # print("prop=%s\nlabels=%s\ndomain_types=%s\nrange_types=%s\n\n" % row)
-        count += 1
 
 subclassMap = {}
 for row in result2:
@@ -107,7 +107,7 @@ for row in result2:
     if currentType in allSchemaOrgTypesDict:
         allSchemaOrgTypesDict[currentType]["parent"] = transform(row, 'subclass')
 
-# Hard coded mandatory props -
+# schema.org does not have any concept of required props, but Google does so we hard code it here
 allSchemaOrgTypesDict["Dataset"][requiredProps] = datasetRequiredProps
 allSchemaOrgTypesDict["DataDownload"][requiredProps] = dataDownloadRequiredProps
 allSchemaOrgTypesDict["Person"][requiredProps] = standardRequiredProps
@@ -116,22 +116,8 @@ allSchemaOrgTypesDict["DataCatalog"][requiredProps] = standardRequiredProps
 
 
 json_object = json.dumps({"schema.org": allSchemaOrgTypesDict}, indent=4)
-f = open("validation.json", "w")
+f = open("jsonldSchema.json", "w")
 f.write(json_object)
 f.close()
 
-# print("count=%d" % (count,))
-result = g.query("""
-SELECT 
-  DISTINCT ?rtype
-WHERE {
-  schema:Dataset rdfs:subClassOf* ?dtype .
-  ?prop a rdf:Property .
-  ?prop schema:domainIncludes ?dtype .
-  ?prop rdfs:label ?label .
-  ?prop schema:rangeIncludes ?rtype .
-}
-""")
 
-# for row in result:
-#    print(row)
