--- test/specs/annotation/conf/google_metadata.conf.json
@@ -1,16 +1,0 @@
-{
-    "catalog": {},
-    "schema": {
-        "name": "google_metadata_schema",
-        "createNew": true,
-        "path": "test/specs/annotation/conf/google_metadata/schema.json"
-    },
-    "tables": {
-        "createNew": true,
-        "name": "google_metadata_w_handlebars"
-    },
-    "entities": {
-        "createNew": true,
-        "path": "test/specs/annotation/conf/google_metadata/data"
-    }
-}
