+++ test/specs/annotation/spec.js
@@ -6,8 +6,7 @@ require('./../../utils/starter.spec.js').runTests({
         "/annotation/tests/04.visible_foreign_keys.js",
         "/annotation/tests/05.citation.js",
         "/annotation/tests/06.catalog.js",
-        "/annotation/tests/07.source_definitions.js",
-        "/annotation/tests/08.google_metadata.js"
+        "/annotation/tests/07.source_definitions.js"
     ],
     schemaConfigurations: [
         "/annotation/conf/displayname.conf.json",
@@ -15,7 +14,6 @@ require('./../../utils/starter.spec.js').runTests({
         "/annotation/conf/visible_foreign_keys.json",
         "/annotation/conf/citation.conf.json",
         "/annotation/conf/catalog.conf.json",
-        "/annotation/conf/source_definitions.conf.json",
-        "/annotation/conf/google_metadata.conf.json"
+        "/annotation/conf/source_definitions.conf.json"
     ]
 });
