+++ test/specs/json-ld/conf/google_metadata/data/google-metadata-w-handlebars.json
