+++ test/support/jasmine.json
@@ -12,7 +12,8 @@
       "print_utils/spec.js",
       "errors/spec.js",
       "upload/spec.js",
-      "export/spec.js"
+      "export/spec.js",
+      "json-ld/spec.js"
   ],
   "helpers": ["../../node_modules/jasmine-expect/index.js"],
   "stopSpecOnExpectationFailure": false,
